64 bit compilation fix
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@1416 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/pdns_recursor.cc
@@ -1955,7 +1955,7 @@ try
   }
   cerr<<"get2 secs: "<<dt.udiff()/1000000.0<<endl;
 #endif 
-  t_id=(int) ptr;
+  t_id=(int) (long) ptr;
   MT=new MTasker<PacketID,string>(::arg().asNum("stack-size"));
